[FLINK-11877][runtime] Abstract AbstractTwoInputStreamTask from TwoInputStreamTask
